ci(workflow): Consolidate CI matrix and resolve mypy & whitespace errors
- Converted the GitHub Actions matrix (.github/workflows/ci.yml) to run a single test suite for Python >=3.10 instead of spawning 3 distinct environments to streamline execution.
- Fixed trailing whitespaces across workflow descriptors.
- Resolved 'floating[Any]' type mismatch inside 'src/syntharc/text/evaluation.py' by applying float() casting.
- Remediated pandas-stubs typing error in 'examples/timeseries/timeseries_generation.py' by using Series.str.cat() instead of the raw '+' operator.
- Pre-commit pipeline successfully passing.

+++ src/syntharc/text/evaluation.py
@@ -84,7 +84,7 @@ def evaluate_text(
     mean_r_len = np.mean(real_lengths)
     mean_s_len = np.mean(synth_lengths)
 
-    len_diff = min(abs(mean_r_len - mean_s_len) / (abs(mean_r_len) + 1e-10), 1.0)
+    len_diff = min(float(abs(mean_r_len - mean_s_len) / (abs(mean_r_len) + 1e-10)), 1.0)
     len_sim = float(1.0 - len_diff)
 
     # 4. Overall score (weight overlap and length similarity)
